requirements.txt
pandas
numpy
pyarrow
seaborn
matplotlib
langchain
langchain-core
langchain-community
langgraph
sentence-transformers
openai
qdrant-client
SQLAlchemy
psycopg2-binary
fastapi
uvicorn
streamlit
python-dotenv
scikit-learn
tqdm
requests